add confirm page
